+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   and even out. Previously the bottom rows were always last to be written.
 - Add more examples: ball, comet, pulser.
 - Refine the keyboard layout grid for GX502.
+- Use tokio to delay main loop, not thread::sleep
 
 ## [0.9.4] - 2020-05-05
 ### Changed